Successfully made the freeze method. Made a debugmove method for playing game on computer for debugging.
Found out the problem with the freeze method. Some of the objects with the "block" tag did not have the "Obstacle" class attached because they were children of untagged objects that did have the "Obstacle" component. In null proofing the program and tagging all the untagged parents, I fixed the errors that the freeze clock was giving me. I also made a method that relied on mouse clicks as opposed to touchcounts so that I could debug the program on my computer. I still have to make sure that the obstacle spawner start time does not restart whenever it is set active, however.